Revisions to provide integrated Rails 3.0 support.
Repertoire Assets now detects when it is running in Rails, and installs
the Rack middleware automatically via a Railtie.  So now the only config
necessary is a single line in production.rb to turn on asset precaching.

The rake tasks have been revised to integrate cleanly with Rails, and
the gemspec configuration follows Rails standards.